+++ src/react_frontend/next.config.js
@@ -1,11 +1,16 @@
+const withPlugins = require('next-with-plugins')
 /** @type {import('next').NextConfig} */
-const nextConfig = {
+const nextConfig = withPlugins({
   reactStrictMode: true,
   swcMinify: true,
   env: {
     NEXT_PUBLIC_BACKEND_URL: 'http://localhost:8000'
   },
-  images: {unoptimized: true}
-}
+  pageFolder: 'src/pages',
+  images: { unoptimized: true },
+  plugins: [
+    'next-stylus'
+  ]
+})
 
 module.exports = nextConfig
