+++ src/react_frontend/package.json
@@ -1,27 +1,34 @@
 {
-  "name": "ckc_project",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "generate": "next build && next export",
-    "lint": "next lint"
-  },
-  "devDependencies": {
-    "@types/node": "18.11.9",
-    "@types/react": "18.0.25",
-    "@types/react-dom": "18.0.9"
-  },
-  "dependencies": {
-    "next": "13.0.5",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "typescript": "4.9.3",
-    "@mui/material": "5.10.16",
-    "@emotion/react": "11.10.5",
-    "@emotion/styled": "11.10.5",
-    "axios": "1.2.0"
-  }
+    "name": "ckc_project",
+    "version": "0.1.0",
+    "private": true,
+    "scripts": {
+        "dev": "next dev",
+        "build": "next build",
+        "start": "next start",
+        "generate": "next build && next export",
+        "lint": "next lint"
+    },
+    "devDependencies": {
+        "@types/node": "18.11.9",
+        "@types/react": "18.0.25",
+        "@types/react-dom": "18.0.9"
+    },
+    "dependencies": {
+        "@emotion/react": "^11.10.5",
+        "@emotion/styled": "^11.10.5",
+        "@mui/icons-material": "^5.11.0",
+        "@mui/material": "^5.11.0",
+        "axios": "1.2.0",
+        "eslint-config-next": "^13.0.7",
+        "material-ui": "^0.20.2",
+        "next": "13.0.5",
+        "next-auth": "^4.18.6",
+        "next-stylus": "^0.0.2",
+        "next-with-plugins": "^0.0.1",
+        "react": "18.2.0",
+        "react-dom": "18.2.0",
+        "stylus": "^0.59.0",
+        "typescript": "4.9.3"
+    }
 }
