+++ src/react_frontend/src/components/layout/header/header.module.css
@@ -1,0 +1,11 @@
+.image {
+    width: 160px;
+    height: 50px;
+}
+
+.link {
+    text-decoration: none;
+    margin-right: 24px;
+    font-size: 24px;
+    color: white;
+}
