+++ src/react_frontend/src/pages/index.tsx
@@ -1,0 +1,16 @@
+import Header from "../components/layout/header/Header";
+
+export default function Home() {
+
+  return (
+    <div>
+      <Header/>
+
+      <main>
+      </main>
+
+      <footer>
+      </footer>
+    </div>
+  )
+}
