+++ src/react_frontend/src/styles/globals.styl
@@ -1,0 +1,38 @@
+@import './globals/flex-helpers'
+@import './globals/font-size'
+@import "./globals/font-styles"
+@import './globals/font-weight'
+@import "./globals/letter-spacing"
+@import './globals/line-height'
+@import './globals/misc'
+@import './globals/opacity'
+@import './globals/overflow'
+@import './globals/position-helpers'
+@import './globals/scrollbars'
+@import './globals/space-helpers'
+@import './globals/z-index-helpers'
+
+html,
+body {
+  padding: 0;
+  margin: 0;
+  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
+    Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
+}
+/*card {*/
+
+/*}*/
+/*a {*/
+/*  color: inherit;*/
+/*  text-decoration: none;*/
+/*}*/
+
+/*@media (prefers-color-scheme: dark) {*/
+/*  html {*/
+/*    color-scheme: dark;*/
+/*  }*/
+/*  body {*/
+/*    color: white;*/
+/*    background: black;*/
+/*  }*/
+/*}*/
