+++ src/react_frontend/src/styles/globals/flex-helpers.styl
@@ -1,0 +1,93 @@
+// Flex container
+.d-flex
+    display: flex
+
+// Flex direction
+.flex-row
+    flex-direction: row
+
+.flex-column
+    flex-direction: column
+
+// Flex wrap
+.flex-wrap
+    flex-wrap: wrap
+
+.flex-no-wrap
+    flex-wrap: nowrap
+
+// Flex grow
+.flex-grow-1
+    flex-grow: 1
+
+.flex-grow-0
+    flex-grow: 0
+
+// Flex shrink
+.flex-shrink-1
+    flex-shrink: 1
+
+.flex-shrink-0
+    flex-shrink: 0
+
+// Flex basis
+.flex-1
+    flex: 1 1 0%
+
+.flex-auto
+    flex: 1 1 auto
+
+// Flex align items
+.flex-start
+    align-items: flex-start
+
+.flex-end
+    align-items: flex-end
+
+.flex-center
+    align-items: center
+
+// Flex align self
+.self-start
+    align-self: flex-start
+
+.self-end
+    align-self: flex-end
+
+.self-center
+    align-self: center
+
+// Flex justify content
+.justify-start
+    justify-content: flex-start
+
+.justify-end
+    justify-content: flex-end
+
+.justify-center
+    justify-content: center
+
+.justify-between
+    justify-content: space-between
+
+.justify-around
+    justify-content: space-around
+
+// Flex order
+.order-1
+    order: 1
+
+.order-2
+    order: 2
+
+.order-3
+    order: 3
+
+.order-4
+    order: 4
+
+.order-5
+    order: 5
+
+.order-6
+    order: 6
