+++ src/react_frontend/src/styles/globals/font-size.styl
@@ -1,0 +1,141 @@
+// Styleguide classes
+.font-11
+    font-size 11px !important
+.font-13
+    font-size 13px !important
+.font-18
+    font-size 18px !important
+.font-21
+    font-size 21px !important
+.font-26
+    font-size 26px !important
+.font-31
+    font-size 31px !important
+.font-39
+    font-size 39px !important
+.font-58
+    font-size 58px !important
+
+// Global Font size classes
+.font-8
+    font-size 8px !important
+
+.font-10
+    font-size 10px !important
+
+.font-12
+    font-size 12px !important
+
+.font-14
+    font-size 14px !important
+
+.font-16
+    font-size 16px !important
+
+.font-18
+    font-size 18px !important
+
+.font-20
+    font-size 20px !important
+
+.font-22
+    font-size 20px !important
+
+.font-24
+    font-size 24px !important
+
+.font-26
+    font-size 26px !important
+
+.font-28
+    font-size 28px !important
+
+.font-30
+    font-size 30px !important
+
+.font-32
+    font-size 32px !important
+
+.font-34
+    font-size 34px !important
+
+.font-36
+    font-size 36px !important
+
+.font-38
+    font-size 38px !important
+
+.font-40
+    font-size 40px !important
+
+.font-42
+    font-size 42px !important
+
+.font-44
+    font-size 44px !important
+
+.font-46
+    font-size 46px !important
+
+.font-48
+    font-size 48px !important
+
+.font-50
+    font-size 50px !important
+
+.font-52
+    font-size 52px !important
+
+.font-54
+    font-size 54px !important
+
+.font-56
+    font-size 56px !important
+
+.font-58
+    font-size 58px !important
+
+.font-60
+    font-size 60px !important
+
+.font-62
+    font-size 62px !important
+
+.font-64
+    font-size 64px !important
+
+.font-66
+    font-size 66px !important
+
+.font-66
+    font-size 66px !important
+
+.font-68
+    font-size 68px !important
+
+.font-70
+    font-size 70px !important
+
+.font-72
+    font-size 72px !important
+
+.font-74
+    font-size 74px !important
+
+.font-76
+    font-size 76px !important
+
+.font-78
+    font-size 78px !important
+
+.font-80
+    font-size 80px !important
+
+.font-82
+    font-size 82px !important
+
+.font-84
+    font-size 84px !important
+
+.font-86
+    font-size 86px !important
