+++ src/react_frontend/src/styles/globals/font-styles.styl
@@ -1,0 +1,5 @@
+.italic
+    font-style italic
+
+.underline
+    text-decoration underline
