+++ src/react_frontend/src/styles/globals/font-weight.styl
@@ -1,0 +1,26 @@
+.weight-100
+    font-weight 100 !important
+
+.weight-200
+    font-weight 200 !important
+
+.weight-300
+    font-weight 300 !important
+
+.weight-400
+    font-weight 400 !important
+
+.weight-500
+    font-weight 500 !important
+
+.weight-600
+    font-weight 600 !important
+
+.weight-700
+    font-weight 700 !important
+
+.weight-800
+    font-weight 800 !important
+
+.weight-900
+    font-weight 900 !important
