+++ src/react_frontend/src/styles/globals/letter-spacing.styl
@@ -1,0 +1,27 @@
+// This file will be built out gradually
+.ls-neg-2
+    letter-spacing -2px !important
+
+.ls-neg-1
+    letter-spacing -1px !important
+
+.ls-0
+    letter-spacing 0px !important
+
+.ls-2
+    letter-spacing 2px !important
+
+.ls-4
+    letter-spacing 4px !important
+
+.ls-6
+    letter-spacing 6px !important
+
+.ls-8
+    letter-spacing 8px !important
+
+.ls-10
+    letter-spacing 10px !important
+
+.ls-12
+    letter-spacing 12px !important
