+++ src/react_frontend/src/styles/globals/line-height.styl
@@ -1,0 +1,118 @@
+// Global Line Height classes
+.lh-2
+    line-height 2px !important
+
+.lh-4
+    line-height 4px !important
+
+.lh-6
+    line-height 6px !important
+
+.lh-8
+    line-height 8px !important
+
+.lh-10
+    line-height 10px !important
+
+
+.lh-12
+    line-height 12px !important
+
+.lh-14
+    line-height 14px !important
+
+.lh-16
+    line-height 16px !important
+
+.lh-18
+    line-height 18px !important
+
+.lh-20
+    line-height 20px !important
+
+.lh-24
+    line-height 24px !important
+
+.lh-26
+    line-height 26px !important
+
+.lh-28
+    line-height 28px !important
+
+.lh-30
+    line-height 30px !important
+
+.lh-32
+    line-height 32px !important
+
+.lh-36
+    line-height 36px !important
+
+.lh-38
+    line-height 38px !important
+
+.lh-40
+    line-height 40px !important
+
+.lh-44
+    line-height 44px !important
+
+.lh-48
+    line-height 48px !important
+
+.lh-50
+    line-height 50px !important
+
+.lh-52
+    line-height 52px !important
+
+.lh-56
+    line-height 56px !important
+
+.lh-58
+    line-height 58px !important
+
+.lh-60
+    line-height 60px !important
+
+.lh-62
+    line-height 62px !important
+
+.lh-64
+    line-height 64px !important
+
+.lh-66
+    line-height 66px !important
+
+.lh-68
+    line-height 68px !important
+
+.lh-68
+    line-height 68px !important
+
+.lh-70
+    line-height 70px !important
+
+.lh-72
+    line-height 72px !important
+
+.lh-74
+    line-height 74px !important
+
+.lh-76
+    line-height 76px !important
+
+.lh-78
+    line-height 78px !important
+
+.lh-80
+    line-height 80px !important
+
+.lh-82
+    line-height 82px !important
+
+.lh-84
+    line-height 84px !important
+
+.lh-86
+    line-height 86px !important
