+++ src/react_frontend/src/styles/globals/misc.styl
@@ -1,0 +1,36 @@
+// Misc
+body, html {
+    margin 0
+    padding 0
+}
+
+.width-container
+    max-width 1280px
+    padding 0 32px
+
+.fill-width
+    width 100%
+
+.flex-1
+    flex 1
+
+.text-nowrap
+    white-space nowrap
+
+.text-underline
+    text-decoration underline
+
+
+.text-none
+    text-decoration none
+
+
+// click event stuff
+.cursor-pointer
+    cursor pointer
+
+.cursor-default
+    cursor default !important
+
+.no-pointer-events
+    pointer-events none
