+++ src/react_frontend/src/styles/globals/overflow.styl
@@ -1,0 +1,20 @@
+.overflow-auto
+    overflow auto
+.overflow-y-auto
+    overflow-y auto
+.overflow-x-auto
+    overflow-x auto
+
+.overflow-hidden
+    overflow hidden
+.overflow-y-hidden
+    overflow-y hidden
+.overflow-x-hidden
+    overflow-x hidden
+
+.overflow-scroll
+    overflow scroll
+.overflow-y-scroll
+    overflow-y scroll
+.overflow-x-scroll
+    overflow-x scroll
