+++ src/react_frontend/src/styles/globals/position-helpers.styl
@@ -1,0 +1,25 @@
+.sticky
+    position sticky !important
+
+.relative
+    position relative !important
+
+.fixed
+    position fixed !important
+
+.absolute
+    position absolute !important
+
+.initial
+    position absolute !important
+.top
+    top 0
+
+.bottom
+    bottom 0
+
+.right
+    right 0
+
+.left
+    left 0
