+++ src/react_frontend/src/styles/globals/space-helpers.styl
@@ -1,0 +1,179 @@
+// Padding
+.pa-0
+  padding: 0
+.pa-1
+  padding: 4px
+.pa-2
+  padding: 8px
+.pa-3
+  padding: 16px
+.pa-4
+  padding: 24px
+.pa-5
+  padding: 32px
+.pa-6
+  padding: 40px
+.pa-7
+  padding: 48px
+
+// Padding top
+.pt-0
+  padding-top: 0
+.pt-1
+  padding-top: 4px
+.pt-2
+  padding-top: 8px
+.pt-3
+  padding-top: 16px
+.pt-4
+  padding-top: 24px
+.pt-5
+  padding-top: 32px
+.pt-6
+  padding-top: 40px
+.pt-7
+  padding-top: 48px
+
+// Padding right
+.pr-0
+  padding-right: 0
+.pr-1
+  padding-right: 4px
+.pr-2
+  padding-right: 8px
+.pr-3
+  padding-right: 16px
+.pr-4
+  padding-right: 24px
+.pr-5
+  padding-right: 32px
+.pr-6
+  padding-right: 40px
+.pr-7
+  padding-right: 48px
+
+// Padding bottom
+.pb-0
+  padding-bottom: 0
+.pb-1
+  padding-bottom: 4px
+.pb-2
+  padding-bottom: 8px
+.pb-3
+  padding-bottom: 16px
+.pb-4
+  padding-bottom: 24px
+.pb-5
+  padding-bottom: 32px
+.pb-6
+  padding-bottom: 40px
+.pb-7
+  padding-bottom: 48px
+
+// Padding left
+.pl-0
+  padding-left: 0
+.pl-1
+  padding-left: 4px
+.pl-2
+  padding-left: 8px
+.pl-3
+  padding-left: 16px
+.pl-4
+  padding-left: 24px
+.pl-5
+  padding-left: 32px
+.pl-6
+  padding-left: 40px
+.pl-7
+  padding-left: 48px
+
+// Margin
+.ma-0
+  margin: 0
+.ma-1
+  margin: 4px
+.ma-2
+  margin: 8px
+.ma-3
+  margin: 16px
+.ma-4
+  margin: 24px
+.ma-5
+  margin: 32px
+.ma-6
+  margin: 40px
+.ma-7
+  margin: 48px
+
+// Margin top
+.mt-0
+  margin-top: 0
+.mt-1
+  margin-top: 4px
+.mt-2
+  margin-top: 8px
+.mt-3
+  margin-top: 16px
+.mt-4
+  margin-top: 24px
+.mt-5
+  margin-top: 32px
+.mt-6
+  margin-top: 40px
+.mt-7
+  margin-top: 48px
+
+// Margin right
+.mr-0
+  margin-right: 0
+.mr-1
+  margin-right: 4px
+.mr-2
+  margin-right: 8px
+.mr-3
+  margin-right: 16px
+.mr-4
+  margin-right: 24px
+.mr-5
+  margin-right: 32px
+.mr-6
+  margin-right: 40px
+.mr-7
+  margin-right: 48px
+
+// Margin bottom
+.mb-0
+  margin-bottom: 0
+.mb-1
+  margin-bottom: 4px
+.mb-2
+  margin-bottom: 8px
+.mb-3
+  margin-bottom: 16px
+.mb-4
+  margin-bottom: 24px
+.mb-5
+  margin-bottom: 32px
+.mb-6
+  margin-bottom: 40px
+.mb-7
+  margin-bottom: 48px
+
+// Margin left
+.ml-0
+  margin-left: 0
+.ml-1
+  margin-left: 4px
+.ml-2
+  margin-left: 8px
+.ml-3
+  margin-left: 16px
+.ml-4
+  margin-left: 24px
+.ml-5
+  margin-left: 32px
+.ml-6
+  margin-left: 40px
+.ml-7
+  margin-left: 48px
