+++ src/react_frontend/src/styles/globals/z-index-helpers.styl
@@ -1,0 +1,33 @@
+
+.z-index-0
+    z-index 0 !important
+
+.z-index-1
+    z-index 1 !important
+
+.z-index-2
+    z-index 2 !important
+
+.z-index-3
+    z-index 3 !important
+
+.z-index-4
+    z-index 4 !important
+
+.z-index-5
+    z-index 5 !important
+
+.z-index-6
+    z-index 6 !important
+
+.z-index-7
+    z-index 7 !important
+
+.z-index-8
+    z-index 8 !important
+
+.z-index-9
+    z-index 9 !important
+
+.z-index-10
+    z-index 10 !important
