+++ src/react_frontend/src/themes/input_theme.js
